Rename //rewriter/utils to //rewriter/util
