+++ rewriter/DSLBuilder.cc
@@ -5,7 +5,7 @@
 #include "core/Names.h"
 #include "core/core.h"
 #include "rewriter/rewriter.h"
-#include "rewriter/utils/Utils.h"
+#include "rewriter/util/Util.h"
 
 using namespace std;
