+++ rewriter/MixinEncryptedProp.cc
@@ -3,7 +3,7 @@
 #include "ast/ast.h"
 #include "core/Context.h"
 #include "core/core.h"
-#include "rewriter/utils/Utils.h"
+#include "rewriter/util/Util.h"
 
 using namespace std;
