+++ rewriter/util/BUILD
@@ -1,7 +1,7 @@
 cc_library(
-    name = "utils",
-    srcs = ["Utils.cc"],
-    hdrs = ["Utils.h"],
+    name = "util",
+    srcs = ["Util.cc"],
+    hdrs = ["Util.h"],
     linkstatic = select({
         "//tools/config:linkshared": 0,
         "//conditions:default": 1,
