+++ rewriter/util/Util.cc
@@ -1,4 +1,4 @@
-#include "rewriter/utils/Utils.h"
+#include "rewriter/util/Util.h"
 #include "absl/strings/escaping.h"
 #include "ast/Helpers.h"
 #include "ast/ast.h"
